\name{longData-internal}
\alias{catShort}
\alias{cleanProg}
\alias{is.tna}
\alias{LETTERSletters}
\alias{meanNA}
\alias{sdNA}

\alias{trajImput.copy.begin}
\alias{trajImput.copy.end}
\alias{trajImput.copy.middle}

\alias{trajImput.copyMean}
\alias{trajImput.globalSlope.begin}
\alias{trajImput.globalSlope.end}
\alias{trajImput.interpoLin}

\alias{trajImput.interpoLin.middle}
\alias{trajImput.interpoLin2}

\alias{trajImput.interpoLin3}
\alias{trajImput.localSlope.begin}

\alias{trajImput.localSlope.end}
\alias{trajImput.LOCB}
\alias{trajImput.LOCB.begin}

\alias{trajImput.LOCB.middle}
\alias{trajImput.LOCF}
\alias{trajImput.LOCF.end}

\alias{trajImput.LOCF.middle}
\alias{which.minNA}


\title{~ longData: internal objects and methods ~}
\description{longData: internal objects and methods}
\details{These are not to be called by the user.\cr
  If you have a specific need or for more details on longData, feel free to contact
  the author
  (which is a hard-worker, will be very happy to answer and - why not? -
  start a new collaboration...)
}


\section{Author(s)}{
  Christophe Genolini\cr
  PSIGIAM: Paris Sud Innovation Group in Adolescent Mental Health\cr
  INSERM U669 / Maison de Solenn / Paris\cr\cr

  Contact author : <genolini@u-paris10.fr>
}

\section{English translation}{
  Raphaël Ricaud\cr
  Laboratoire "Sport & Culture" / "Sports & Culture" Laboratory \cr
  University of Paris 10 / Nanterre
}

\keyword{internal}

